Hierarchy roles updated. AddAdmin method moved to ClientController
