+++ src/Controller/UserController.php
@@ -30,7 +30,7 @@ class UserController extends FOSRestController
      *     requirements={"id"="\d+"}
      * )
      * @return JsonResponse|Response
-     * @Security("is_granted('ROLE_ADMIN')")
+     * @Security("is_granted(['ROLE_ADMIN', 'ROLE_SUPER_ADMIN'])")
      */
     public function showAction(User $user, Request $request, UserManager $userManager)
     {
@@ -47,7 +47,7 @@ class UserController extends FOSRestController
      *     path="/api/users",
      *     name="user_list_all"
      * )
-     * @Security("is_granted('ROLE_ADMIN')")
+     * @Security("is_granted(['ROLE_ADMIN', 'ROLE_SUPER_ADMIN'])")
      */
     public function listAction(Request $request, UserManager $userManager)
     {
@@ -73,25 +73,6 @@ class UserController extends FOSRestController
             return new JsonResponse($data, 400);
         }
         return $this->generateCustomView($data, 201, 'user_add');
-    }
-
-    /**
-     * @param Request $request
-     * @param UserManager $userManager
-     * @Rest\Post(
-     *     path="/api/clients/{id}/admin",
-     *     name="admin_add",
-     *     requirements={"id"="\d+"}
-     * )
-     * @Security("is_granted('ROLE_SUPER_ADMIN')")
-     */
-    public function createAdminAction(Request $request, UserManager $userManager, $id)
-    {
-        $data = $userManager->registerAdmin($request, $id);
-        if (is_array($data)) {
-            return new JsonResponse($data, 400);
-        }
-        return $this->generateCustomView($data, 201, 'admin_add', ['id' => $id]);
     }
 
     /**
